refactor run monitor into modular hooks
